refactor: P2-T1 - 从CalibrationStepView提取CalibrationStageCard等4个类型到Views/Calibration/
