+++ LonelyPianistAVP/Views/Calibration/CalibrationSimulatorDemoState.swift
@@ -1,0 +1,3 @@
+enum CalibrationSimulatorDemoState: Hashable {
+    case enabled
+}
